requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
attrs==23.2.0
bitarray==2.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.0
cytoolz==0.12.3
eth-account==0.11.1
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.1.0
eth-utils==4.1.0
eth_abi==5.1.0
frozenlist==1.4.1
greenlet==3.0.3
hexbytes==0.3.1
idna==3.6
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
parsimonious==0.10.0
pluggy==1.4.0
protobuf==5.26.1
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pytest==8.1.1
python-dotenv==1.0.1
pyunormalize==15.1.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.18.0
SQLAlchemy==2.0.29
toolz==0.12.1
typing_extensions==4.10.0
urllib3==2.2.1
web3==6.16.0
websockets==12.0
yarl==1.9.4
